Make profile caching work for all users (Current and Visitor)
Refactored the loadProfileContent and moved it to ProfilePage, and made it use abstract
method loadUserProfile which is implemented differently in each profile page type.
Also added some robustness to FirebaseCloudStorage in case the userEmail is null.
